requirements-docs.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --extra=docs --output-file=requirements-docs.txt pyproject.toml
babel==2.15.0
    # via
    #   mkdocs-git-revision-date-localized-plugin
    #   mkdocs-material
boto3==1.34.136
    # via mlinfra (pyproject.toml)
botocore==1.34.136
    # via
    #   boto3
    #   s3transfer
certifi==2024.8.30
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via mkdocs
colorama==0.4.6
    # via mkdocs-material
getmac==0.9.5
    # via mlinfra (pyproject.toml)
ghp-import==2.1.0
    # via mkdocs
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via
    #   mlinfra (pyproject.toml)
    #   mkdocs-git-revision-date-localized-plugin
idna==3.10
    # via requests
importlib-metadata==8.0.0
    # via mike
importlib-resources==6.4.0
    # via mike
invoke==2.2.0
    # via mlinfra (pyproject.toml)
jinja2==3.1.4
    # via
    #   mike
    #   mkdocs
    #   mkdocs-material
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
markdown==3.6
    # via
    #   mkdocs
    #   mkdocs-material
    #   pymdown-extensions
markupsafe==2.1.5
    # via
    #   jinja2
    #   mkdocs
mergedeep==1.3.4
    # via
    #   mkdocs
    #   mkdocs-get-deps
mike==2.1.2
    # via mlinfra (pyproject.toml)
mkdocs==1.6.0
    # via
    #   mike
    #   mkdocs-git-revision-date-localized-plugin
    #   mkdocs-material
    #   mkdocs-section-index
mkdocs-get-deps==0.2.0
    # via mkdocs
mkdocs-git-revision-date-localized-plugin==1.2.6
    # via mlinfra (pyproject.toml)
mkdocs-material==9.5.27
    # via mlinfra (pyproject.toml)
mkdocs-material-extensions==1.3.1
    # via mkdocs-material
mkdocs-section-index==0.3.9
    # via mlinfra (pyproject.toml)
mypy==1.12.0
    # via mlinfra (pyproject.toml)
mypy-extensions==1.0.0
    # via mypy
packaging==24.1
    # via mkdocs
paginate==0.5.7
    # via mkdocs-material
pathspec==0.12.1
    # via mkdocs
platformdirs==4.3.6
    # via mkdocs-get-deps
pygments==2.18.0
    # via mkdocs-material
pymdown-extensions==10.8.1
    # via mkdocs-material
pyparsing==3.2.0
    # via mike
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   ghp-import
pytz==2024.1
    # via mkdocs-git-revision-date-localized-plugin
pyyaml==6.0.2
    # via
    #   mlinfra (pyproject.toml)
    #   mike
    #   mkdocs
    #   mkdocs-get-deps
    #   pymdown-extensions
    #   pyyaml-env-tag
pyyaml-env-tag==0.1
    # via
    #   mike
    #   mkdocs
regex==2024.5.15
    # via mkdocs-material
requests==2.32.3
    # via
    #   mlinfra (pyproject.toml)
    #   mkdocs-material
s3transfer==0.10.3
    # via boto3
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
typing-extensions==4.12.2
    # via mypy
urllib3==2.2.2
    # via
    #   botocore
    #   requests
verspec==0.1.0
    # via mike
watchdog==4.0.1
    # via mkdocs
zipp==3.19.2
    # via importlib-metadata
